Colocate erdos-renyi-graph, moving its shared surface into viz/flow
The systems-biology series rendered the same explorer and reached into
this block for the generator hook and getEdgeOptions, which the layer
boundary now rejects. Two pieces render it, so it is no longer private to
either: the explorer, the hook and the G(n,M) generation all move down.

None of the three is in the viz/flow barrel. Naming them there put 1.2 kB
on /blocks/c1-ffl and /blocks/circuit-evolution, neither of which draws a
random graph, and the explorer would have added katex besides.
