+++ pages/blocks/erdos-renyi-graph/figures/index.ts
@@ -1,0 +1,2 @@
+// The explorer lives in viz/ because the systems-biology series renders it too.
+export { default as ErdosRenyiGNMNetwork } from "@viz/flow/RandomGraphExplorer";
