+++ pages/blocks/erdos-renyi-graph/meta.ts
@@ -1,0 +1,13 @@
+import { definePiece } from "@lib/piece";
+
+export default definePiece({
+  href: "/blocks/erdos-renyi-graph",
+  kind: "block",
+  title: "Erdős-Rényi Graphs",
+  subtitle: "How to model random networks",
+  description:
+    "Generate Erdős-Rényi random graphs in the browser: set the number of nodes and edges in the G(n, M) model and watch the network rewire as you change them.",
+  keywords: "Erdős-Rényi graphs, random networks, graph theory, probability",
+  published: "2024-10-30",
+  art: "network",
+});
